Refactor EA collection cards and sequence cards into their own files
