+++ packages/lesswrong/components/ea-forum/EABestOfPage.tsx
@@ -1,14 +1,11 @@
 import React, { useMemo, useRef } from "react";
-import { Components, registerComponent, slugify } from "../../lib/vulcan-lib";
+import { Components, registerComponent } from "../../lib/vulcan-lib";
 import classNames from "classnames";
 import { siteImageSetting } from "../vulcan-core/App";
 import moment from "moment";
 import { InteractionWrapper } from "../common/useClickableCell";
 import { postGetPageUrl } from "../../lib/collections/posts/helpers";
 import { Link } from "../../lib/reactRouterWrapper";
-import { isEAForum } from "../../lib/instanceSettings";
-import { sequenceGetPageUrl } from "../../lib/collections/sequences/helpers";
-import { collectionGetPageUrl } from "../../lib/collections/collections/helpers";
 import { AnalyticsContext } from "../../lib/analyticsEvents";
 import { useHover } from "../common/withHover";
 import { useMulti } from "../../lib/crud/withMulti";
@@ -16,11 +13,6 @@ import keyBy from "lodash/keyBy";
 
 const MAX_WIDTH = 1500;
 const MD_WIDTH = 1000;
-
-// For overlay on sequence and collection cards
-const SEQUENCE_CARD_IMAGE_HEIGHT = 162;
-const Z_IMAGE = 1;
-const Z_OVERLAY = 2;
 
 const styles = (theme: ThemeType): JssStyles => ({
   root: {
@@ -154,58 +146,6 @@ const styles = (theme: ThemeType): JssStyles => ({
       display: "none",
     },
   },
-  // Sequence or collection card
-  sequenceCard: {
-    borderRadius: theme.borderRadius.default,
-    background: theme.palette.panelBackground.default,
-    fontFamily: theme.palette.fonts.sansSerifStack,
-    fontWeight: 500,
-    marginLeft: 8, // Account for box shadow
-    boxShadow: `/* The top layer shadow */
-      -1px 1px 10px ${theme.palette.greyAlpha(0.15)},
-      /* The second layer */
-      -8px 8px 0 0px ${theme.palette.panelBackground.default},
-      /* The second layer shadow */
-      -10px 10px 10px -1px ${theme.palette.greyAlpha(0.15)}`,
-  },
-  sequenceCardImageWrapper: {
-    position: "relative",
-  },
-  sequenceCardImage: {
-    width: "100%",
-    height: SEQUENCE_CARD_IMAGE_HEIGHT,
-    objectFit: "cover",
-    borderTopLeftRadius: theme.borderRadius.default,
-    borderTopRightRadius: theme.borderRadius.default,
-    zIndex: Z_IMAGE,
-  },
-  sequenceReadProgress: {
-    position: "absolute",
-    zIndex: Z_OVERLAY,
-    fontSize: 12,
-    top: 8,
-    left: 8,
-    borderRadius: 14,
-    backgroundColor: theme.palette.panelBackground.translucent3,
-    padding: "6px 8px",
-  },
-  sequenceCardText: {
-    padding: 16,
-  },
-  sequenceCardTitle: {
-    fontSize: 18,
-    lineHeight: "20px",
-    overflow: "hidden",
-    display: "-webkit-box",
-    "-webkit-box-orient": "vertical",
-    "-webkit-line-clamp": 2,
-    marginBottom: 6,
-  },
-  sequenceCardMeta: {
-    fontSize: 14,
-    lineHeight: "20px",
-    color: theme.palette.grey[600],
-  },
   // Featured audio card
   audioCard: {
     marginTop: 16,
@@ -325,119 +265,6 @@ const PostListItem = ({
   );
 };
 
-const SequenceOrCollectionCard = ({
-  title,
-  author,
-  postCount,
-  readCount,
-  imageId,
-  href,
-  eventHandlers,
-  classes,
-}: {
-  title: string;
-  author: UsersMinimumInfo | null;
-  postCount: number;
-  readCount: number;
-  imageId: string;
-  href: string;
-  eventHandlers: {
-    onMouseOver: (event: AnyBecauseTodo) => void;
-    onMouseLeave: () => void;
-  };
-  classes: ClassesType;
-}) => {
-  const { CloudinaryImage2, UsersNameDisplay } = Components;
-
-  const readProgress = `${readCount}/${postCount}`;
-
-  return (
-    <div {...eventHandlers} className={classes.sequenceCard}>
-      <div className={classes.sequenceCardImageWrapper}>
-        <CloudinaryImage2 publicId={imageId} className={classes.sequenceCardImage} />
-        <div className={classes.sequenceReadProgress}>{readProgress}</div>
-      </div>
-      <div className={classes.sequenceCardText}>
-        <Link to={href} className={classes.sequenceCardTitle}>
-          {title}
-        </Link>
-        <div className={classes.sequenceCardMeta}>
-          <UsersNameDisplay user={author} />
-          {" · "}
-          {postCount} posts
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const CollectionCard = ({ collection, classes }: { collection: CollectionsBestOfFragment; classes: ClassesType }) => {
-  const { eventHandlers } = useHover({
-    pageElementContext: "collectionCard",
-    documentId: collection._id,
-    documentSlug: collection.slug,
-  });
-
-  const title = collection.title;
-  const author = collection.user;
-
-  const imageId =
-    // TODO JP-look-here
-    collection.gridImageId || (isEAForum ? "Banner/yeldubyolqpl3vqqy0m6.jpg" : "sequences/vnyzzznenju0hzdv6pqb.jpg");
-  const href = collectionGetPageUrl(collection);
-
-  return (
-    <AnalyticsContext documentSlug={collection?.slug ?? "unknown-slug"}>
-      <SequenceOrCollectionCard
-        title={title}
-        author={author}
-        postCount={collection.postsCount}
-        readCount={collection.readPostsCount}
-        imageId={imageId}
-        href={href}
-        eventHandlers={eventHandlers}
-        classes={classes}
-      />
-    </AnalyticsContext>
-  );
-};
-
-const SequenceCard = ({ sequence, classes }: { sequence: SequencesPageFragment; classes: ClassesType }) => {
-  // Note: this is not a real slug, it's just so we can recognise the sequence in the analytics,
-  // without risking any weirdness due to titles having spaces in them
-  const slug = slugify(sequence?.title ?? "unknown-slug");
-
-  const { eventHandlers } = useHover({
-    pageElementContext: "sequenceCard",
-    documentId: sequence._id,
-    documentSlug: slug,
-  });
-
-  const title = sequence.title;
-  const author = sequence.user;
-
-  const imageId =
-    sequence.gridImageId ||
-    sequence.bannerImageId ||
-    (isEAForum ? "Banner/yeldubyolqpl3vqqy0m6.jpg" : "sequences/vnyzzznenju0hzdv6pqb.jpg");
-  const href = sequenceGetPageUrl(sequence);
-
-  return (
-    <AnalyticsContext documentSlug={slug}>
-      <SequenceOrCollectionCard
-        title={title}
-        author={author}
-        postCount={sequence.postsCount}
-        readCount={sequence.readPostsCount}
-        imageId={imageId}
-        href={href}
-        eventHandlers={eventHandlers}
-        classes={classes}
-      />
-    </AnalyticsContext>
-  );
-};
-
 const AudioPostCard = ({ post, classes }: { post: PostsBestOfList; classes: ClassesType }) => {
   const { PostsPodcastPlayer } = Components;
 
@@ -459,7 +286,7 @@ const AudioPostCard = ({ post, classes }: { post: PostsBestOfList; classes: Clas
 };
 
 const EABestOfPage = ({ classes }: { classes: ClassesType }) => {
-  const { HeadTags } = Components;
+  const { HeadTags, EASequenceCard, EACollectionCard } = Components;
 
   const { results: posts, loading } = useMulti({
     terms: {postIds: allPostIds, limit: allPostIds.length},
@@ -504,7 +331,7 @@ const EABestOfPage = ({ classes }: { classes: ClassesType }) => {
                 <h2 className={classes.heading}>Featured collections</h2>
                 <div className={classes.gridSection}>
                   {featuredCollectionSequences.map((sequence) => (
-                    <SequenceCard key={sequence._id} sequence={sequence} classes={classes} />
+                    <EASequenceCard key={sequence._id} sequence={sequence} />
                   ))}
                 </div>
               </div>
@@ -524,10 +351,10 @@ const EABestOfPage = ({ classes }: { classes: ClassesType }) => {
                 <h2 className={classes.heading}>Learn about Effective Altruism</h2>
                 <div className={classes.gridSection}>
                   {learnAboutEACollections.map((collection) => (
-                    <CollectionCard key={collection._id} collection={collection} classes={classes} />
+                    <EACollectionCard key={collection._id} collection={collection} />
                   ))}
                   {learnAboutEASequences.map((sequence) => (
-                    <SequenceCard key={sequence._id} sequence={sequence} classes={classes} />
+                    <EASequenceCard key={sequence._id} sequence={sequence} />
                   ))}
                 </div>
               </div>
@@ -537,7 +364,7 @@ const EABestOfPage = ({ classes }: { classes: ClassesType }) => {
                 <h2 className={classes.heading}>Intro to cause areas</h2>
                 <div className={classes.gridSection}>
                   {introToCauseAreasSequences.map((sequence) => (
-                    <SequenceCard key={sequence._id} sequence={sequence} classes={classes} />
+                    <EASequenceCard key={sequence._id} sequence={sequence} />
                   ))}
                 </div>
               </div>
