+++ packages/lesswrong/components/ea-forum/EASequenceCard.tsx
@@ -1,0 +1,54 @@
+import React from "react";
+import { Components, registerComponent, slugify } from "../../lib/vulcan-lib";
+import { useHover } from "../common/withHover";
+import { isEAForum } from "../../lib/instanceSettings";
+import { sequenceGetPageUrl } from "../../lib/collections/sequences/helpers";
+import { AnalyticsContext } from "../../lib/analyticsEvents";
+
+const EASequenceCard = ({sequence}: {sequence: SequencesPageFragment}) => {
+  // Note: this is not a real slug, it's just so we can recognise the sequence
+  // in the analytics, without risking any weirdness due to titles having spaces 
+  // in them.
+  const slug = slugify(sequence?.title ?? "unknown-slug");
+
+  const {eventHandlers} = useHover({
+    pageElementContext: "sequenceCard",
+    documentId: sequence._id,
+    documentSlug: slug,
+  });
+
+  const title = sequence.title;
+  const author = sequence.user;
+
+  const imageId =
+    sequence.gridImageId ||
+    sequence.bannerImageId ||
+    (isEAForum ? "Banner/yeldubyolqpl3vqqy0m6.jpg" : "sequences/vnyzzznenju0hzdv6pqb.jpg");
+  const href = sequenceGetPageUrl(sequence);
+
+  const {EASequenceOrCollectionCard} = Components;
+  return (
+    <AnalyticsContext documentSlug={slug}>
+      <EASequenceOrCollectionCard
+        title={title}
+        author={author}
+        postCount={sequence.postsCount}
+        readCount={sequence.readPostsCount}
+        imageId={imageId}
+        href={href}
+        eventHandlers={eventHandlers}
+      />
+    </AnalyticsContext>
+  );
+};
+
+const EASequenceCardComponent = registerComponent(
+  "EASequenceCard",
+  EASequenceCard,
+);
+
+declare global {
+  interface ComponentTypes {
+    EASequenceCard: typeof EASequenceCardComponent;
+  }
+}
